In PluginHost Build Phases, two Copy Files phases were renamed to Copy Frameworks and Copy Plugins.
Also in PluginHost Build Phases, PluginTestA was added as a target dependency of PluginHost.  This addition should provide automatic code signing for the plugin when Xcode builds PluginHost.  New plugins should be added to Target Dependencies and Copy Plugins phases in order to embed the plugins in the application bundle.